Set workflow id required. Making workflow id required with format checks done at the swadl-schema level allows us to make sure when a workflow changed, it has an id and then the old version is not used. Like this we do not need to handle cases where id is not present and how to generate one.
