+++ workflow-bot-app/src/main/java/com/symphony/bdk/workflow/swadl/SwadlParser.java
@@ -47,13 +47,6 @@ public class SwadlParser {
   public static Workflow fromYaml(File workflowFile) throws IOException, ProcessingException {
     String yamlString = Files.readString(workflowFile.toPath(), StandardCharsets.UTF_8);
     SwadlValidator.validateYaml(yamlString);
-    Workflow workflow = MAPPER.readValue(yamlString, Workflow.class);
-    if (workflow.getId() == null) {
-      workflow.setId(workflowFile.getName()
-          .replaceAll(".swadl", "")
-          .replaceAll(".yaml", ""));
-    }
-
-    return workflow;
+    return MAPPER.readValue(yamlString, Workflow.class);
   }
 }
