+++ workflow-bot-app/src/test/java/com/symphony/bdk/workflow/RoomIntegrationTest.java
@@ -46,7 +46,7 @@ class RoomIntegrationTest extends IntegrationTest {
     stream.setId("0000");
     when(streamService.create(uids)).thenReturn(stream);
 
-    engine.deploy(workflow, "defaultId");
+    engine.deploy(workflow);
     engine.onEvent(messageReceived("/create-mim"));
     verify(streamService, timeout(5000)).create(uids);
   }
@@ -77,7 +77,7 @@ class RoomIntegrationTest extends IntegrationTest {
             crossPod, copyProtected, multilateralRoom, memberCanInvite, subType, keywords);
     when(streamService.create(any(V3RoomAttributes.class))).thenReturn(v3RoomDetail);
 
-    engine.deploy(workflow, "defaultId");
+    engine.deploy(workflow);
     engine.onEvent(IntegrationTest.messageReceived("/create-room"));
 
     ArgumentCaptor<V3RoomAttributes> argumentCaptor = ArgumentCaptor.forClass(V3RoomAttributes.class);
@@ -107,7 +107,7 @@ class RoomIntegrationTest extends IntegrationTest {
             null, null);
     when(streamService.create(any(V3RoomAttributes.class))).thenReturn(v3RoomDetail);
 
-    engine.deploy(workflow, "defaultId");
+    engine.deploy(workflow);
     engine.onEvent(IntegrationTest.messageReceived("/create-room-members"));
 
     ArgumentCaptor<V3RoomAttributes> argumentCaptor = ArgumentCaptor.forClass(V3RoomAttributes.class);
@@ -198,7 +198,7 @@ class RoomIntegrationTest extends IntegrationTest {
     roomDetail.setRoomSystemInfo(info);
     when(streamService.getRoomInfo("abc")).thenReturn(roomDetail);
 
-    engine.deploy(workflow, "defaultId");
+    engine.deploy(workflow);
     engine.onEvent(messageReceived("/update-room"));
 
     ArgumentCaptor<V3RoomAttributes> attributes = ArgumentCaptor.forClass(V3RoomAttributes.class);
@@ -223,7 +223,7 @@ class RoomIntegrationTest extends IntegrationTest {
     roomDetail.setRoomSystemInfo(info);
     when(streamService.getRoomInfo("abc")).thenReturn(roomDetail);
 
-    engine.deploy(workflow, "defaultId");
+    engine.deploy(workflow);
     engine.onEvent(messageReceived("/update-room"));
 
     verify(streamService, timeout(5000)).setRoomActive("abc", true);
@@ -234,7 +234,7 @@ class RoomIntegrationTest extends IntegrationTest {
     final Workflow workflow =
         SwadlParser.fromYaml(getClass().getResourceAsStream("/room/add-room-member.swadl.yaml"));
 
-    engine.deploy(workflow, "defaultId");
+    engine.deploy(workflow);
     engine.onEvent(messageReceived("/add-room-member"));
 
     verify(streamService, timeout(5000)).addMemberToRoom(123L, "abc");
@@ -246,7 +246,7 @@ class RoomIntegrationTest extends IntegrationTest {
     final Workflow workflow =
         SwadlParser.fromYaml(getClass().getResourceAsStream("/room/remove-room-member.swadl.yaml"));
 
-    engine.deploy(workflow, "defaultId");
+    engine.deploy(workflow);
     engine.onEvent(messageReceived("/remove-room-member"));
 
     verify(streamService, timeout(5000)).removeMemberFromRoom(123L, "abc");
@@ -258,7 +258,7 @@ class RoomIntegrationTest extends IntegrationTest {
     final Workflow workflow =
         SwadlParser.fromYaml(getClass().getResourceAsStream("/room/promote-room-owner.swadl.yaml"));
 
-    engine.deploy(workflow, "defaultId");
+    engine.deploy(workflow);
     engine.onEvent(messageReceived("/promote-room-owner"));
 
     verify(streamService, timeout(5000)).promoteUserToRoomOwner(123L, "abc");
@@ -270,7 +270,7 @@ class RoomIntegrationTest extends IntegrationTest {
     final Workflow workflow =
         SwadlParser.fromYaml(getClass().getResourceAsStream("/room/demote-room-owner.swadl.yaml"));
 
-    engine.deploy(workflow, "defaultId");
+    engine.deploy(workflow);
     engine.onEvent(messageReceived("/demote-room-owner"));
 
     verify(streamService, timeout(5000)).demoteUserToRoomParticipant(123L, "abc");
@@ -282,7 +282,7 @@ class RoomIntegrationTest extends IntegrationTest {
     final Workflow workflow = SwadlParser.fromYaml(getClass().getResourceAsStream("/room/get-room.swadl.yaml"));
     when(streamService.getRoomInfo("abc")).thenReturn(new V3RoomDetail());
 
-    engine.deploy(workflow, "defaultId");
+    engine.deploy(workflow);
     engine.onEvent(messageReceived("/get-room"));
 
     verify(streamService, timeout(5000)).getRoomInfo("abc");
@@ -295,7 +295,7 @@ class RoomIntegrationTest extends IntegrationTest {
         SwadlParser.fromYaml(getClass().getResourceAsStream("/room/get-room-members.swadl.yaml"));
     when(streamService.listRoomMembers("abc")).thenReturn(Collections.emptyList());
 
-    engine.deploy(workflow, "defaultId");
+    engine.deploy(workflow);
     engine.onEvent(messageReceived("/get-room-members"));
 
     verify(streamService, timeout(5000)).listRoomMembers("abc");
@@ -314,7 +314,7 @@ class RoomIntegrationTest extends IntegrationTest {
         .sortOrder(V2RoomSearchCriteria.SortOrderEnum.BASIC);
     when(streamService.searchRooms(refEq(query))).thenReturn(new V3RoomSearchResults());
 
-    engine.deploy(workflow, "defaultId");
+    engine.deploy(workflow);
     engine.onEvent(messageReceived("/get-rooms"));
 
     verify(streamService, timeout(5000)).searchRooms(refEq(query));
@@ -334,7 +334,7 @@ class RoomIntegrationTest extends IntegrationTest {
     when(streamService.searchRooms(refEq(query), refEq(new PaginationAttribute(10, 10))))
         .thenReturn(new V3RoomSearchResults());
 
-    engine.deploy(workflow, "defaultId");
+    engine.deploy(workflow);
     engine.onEvent(messageReceived("/get-rooms-pagination"));
 
     verify(streamService, timeout(5000)).searchRooms(refEq(query), refEq(new PaginationAttribute(10, 10)));
