+++ workflow-bot-app/src/test/resources/branching/if.swadl.yaml
@@ -1,3 +1,4 @@
+id: branching-workflow
 variables:
   foo: foo
 activities:
