+++ workflow-bot-app/src/test/resources/deployment/deployment-different-workflow-same-id-2.swadl.yaml
@@ -1,3 +1,4 @@
+id: same_workflow-id
 activities:
   - send-message:
       id: act1
