+++ workflow-language/src/main/java/com/symphony/bdk/workflow/swadl/v1/Workflow.java
@@ -7,12 +7,11 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import javax.annotation.Nullable;
 
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Workflow {
-  @Nullable private String id;
+  private String id;
   private Map<String, Object> variables = Collections.emptyMap();
   private List<Activity> activities;
