+++ workflow-language/src/main/resources/swadl-schema-1.0.json
@@ -5,13 +5,14 @@
     "type": "object",
     "additionalProperties": false,
     "required": [
+        "id",
         "activities"
     ],
     "properties": {
         "id": {
             "type": "string",
             "description": "Workflow's id should start with a letter. It is optional. If no id is provided, the workflow filename is used. Otherwise, it will be generated. The id will appear in logs and audit trails.",
-            "pattern": "^[a-zA-Z]",
+            "pattern": "^[a-zA-Z]+[a-zA-Z0-9][^ ]*$",
             "default": "my-workflow"
         },
         "variables": {
